+++ src/command_modules/azure-cli-role/azure/cli/command_modules/role/custom.py
@@ -606,7 +606,7 @@ def create_application(client, display_name, homepage=None, identifier_uris=None
 def update_application(client, identifier, display_name=None, homepage=None,
                        identifier_uris=None, password=None, reply_urls=None, key_value=None,
                        key_type=None, key_usage=None, start_date=None, end_date=None, available_to_other_tenants=None,
-                       oauth2_allow_implicit_flow=None, required_resource_accesses=None):
+                       oauth2_allow_implicit_flow=None, required_resource_accesses=None, additional_properties=None):
     object_id = _resolve_application(client, identifier)
 
     password_creds, key_creds, required_accesses = None, None, None
@@ -626,6 +626,19 @@ def update_application(client, identifier, display_name=None, homepage=None,
                                                   available_to_other_tenants=available_to_other_tenants,
                                                   required_resource_access=required_accesses,
                                                   oauth2_allow_implicit_flow=oauth2_allow_implicit_flow)
+    if additional_properties:
+        app_patch_param.enable_additional_properties_sending()
+        if app_patch_param.additional_properties is None:
+            app_patch_param.additional_properties = {}
+
+        for p in additional_properties:
+            name, value = p.split('=')
+            try:
+                value = shell_safe_json_parse(value)  # see whether it is a json
+            except CLIError:
+                pass
+            app_patch_param.additional_properties[name] = value
+
     return client.patch(object_id, app_patch_param)