+++ src/command_modules/azure-cli-role/azure/cli/command_modules/role/tests/latest/test_graph.py
@@ -110,6 +110,11 @@ class ApplicationSetScenarioTest(ScenarioTest):
         self.cmd('ad app show --id {app}',
                  checks=self.check('replyUrls[0]', '{reply_uri}'))
 
+        # update through generic --additional-properties
+        self.cmd('ad app update --id {app} --additional-properties oauth2AllowUrlPathMatching=true')
+        self.cmd('ad app show --id {app}',
+                 checks=self.check('oauth2AllowUrlPathMatching', True))
+
         # delete app
         self.cmd('ad app delete --id {app}')
         self.cmd('ad app list --identifier-uri {app}',